Extract isSubset to own file
And use  polyfill, as IE11 doesn't support
