+++ lib/samsam.js
@@ -9,28 +9,7 @@ var isElement = require("./is-element");
 var isNaN = require("./is-nan");
 var isNegZero = require("./is-neg-zero");
 var isSet = require("./is-set");
-
-function isSubset(s1, s2, compare) {
-    var values1 = Array.from(s1);
-    var values2 = Array.from(s2);
-
-    for (var i = 0; i < values1.length; i++) {
-        var includes = false;
-
-        for (var j = 0; j < values2.length; j++) {
-            if (compare(values2[j], values1[i])) {
-                includes = true;
-                break;
-            }
-        }
-
-        if (!includes) {
-            return false;
-        }
-    }
-
-    return true;
-}
+var isSubset = require("./is-subset");
 
 /**
  * @name samsam.deepEqual
